feat: 整理example工程，分页面显示demo
