+++ examples/mini-program-example/src/app.config.ts
@@ -4,7 +4,11 @@ export default defineAppConfig({
     'pages/grid-view/index',
     'pages/list-view/index',
     'pages/page-container/index',
-    'pages/match-media/index'
+    'pages/match-media/index',
+    'pages/sticky-header/index',
+    'pages/root-portal/index',
+    'pages/radio-group/index',
+    'pages/check-box-group/index',
   ],
   window: {
     backgroundTextStyle: 'light',
