+++ examples/mini-program-example/src/pages/radio-group/index.config.ts
@@ -1,0 +1,3 @@
+export default definePageConfig({
+  navigationBarTitleText: '第二个页面'
+})
