+++ examples/mini-program-example/src/pages/sticky-header/index.scss
@@ -20,6 +20,3 @@
 
 
 
-.buttonStyle {
-  width: 200px;
-}
