+++ packages/taro-components-mpharmony/src/components/index.ts
@@ -4,4 +4,6 @@ export * from './label/label'
 export * from './list/list'
 export * from './page-container/page-container'
 export * from './root-portal/root-portal'
+export * from './sticky-header/sticky-header'
+export * from './sticky-section/sticky-section'
